Finished cart display, time to the purchase button.
In this commit i added the remove, and update amount button on the cart.
I also added some validations to the purchase button, if there is no items on
the cart then the user will not be able to click the button. If the user is not
logged in and there are items on the cart, then the button will redirect the
user to the login page.